djlfs_batch/requirements.txt
Django==1.11.5
PyJWT==1.5.3
